feat: separate documentation for mkdocs
It was always the intent to use MKDocs to generate descriptive
documentation around this repository. As the content of the README.md
file grew, it became clear that most of the docs should be moved to the
docs directory. This commit does that, and cleans up a lot of the docs
in the process.

* Moved a lot of documentation from the readme to docs/
* Revamped the .devcontainer file to support docker-in-docker